+++ packages/hierarchies/src/test/HierarchyFiltering.test.ts
@@ -11,10 +11,11 @@ describe("HierarchyFilteringPath", () => {
     describe("autoExpand", () => {
       const optionsInOrderOfPriority: Array<HierarchyFilteringPathOptions | undefined> = [
         { autoExpand: true },
-        { autoExpand: { depth: 2 } },
+        { autoExpand: { depth: 3 } },
+        { autoExpand: { depthInPath: 2 } },
         { autoExpand: { depth: 1 } },
-        { autoExpand: { depth: 4, includeGroupingNodes: true } },
-        { autoExpand: { depth: 3, includeGroupingNodes: true } },
+        { autoExpand: { depth: 6, includeGroupingNodes: true } },
+        { autoExpand: { depthInHierarchy: 5 } },
         { autoExpand: false },
         undefined,
       ];