+++ packages/hierarchies/src/test/imodel/FilteringHierarchyDefinition.test.ts
@@ -337,14 +337,14 @@ describe("FilteringHierarchyDefinition", () => {
       expect(node.autoExpand).to.be.undefined;
     });
 
-    it("doesn't set auto-expand when all filtered children paths autoExpand depth is equal to parent keys length", async () => {
+    it("doesn't set auto-expand when all filtered children paths autoExpand depthInPath is equal to parent keys length", async () => {
       const paths: HierarchyFilteringPath[] = [
         {
           path: [
             createTestInstanceKey({ id: "0x1", className: "TestSchema.TestName" }),
             createTestInstanceKey({ id: "0x2", className: "TestSchema.TestName" }),
           ],
-          options: { autoExpand: { depth: 1 } },
+          options: { autoExpand: { depthInPath: 1 } },
         },
       ];
       const filteringFactory = await createFilteringHierarchyDefinition({ nodeIdentifierPaths: paths });
@@ -369,6 +369,26 @@ describe("FilteringHierarchyDefinition", () => {
           },
         }),
       );
+      expect(node.autoExpand).to.be.undefined;
+    });
+
+    it("doesn't set auto-expand when all filtered children paths autoExpand depthInPath is equal to node depth", async () => {
+      const paths: HierarchyFilteringPath[] = [
+        {
+          path: [
+            createTestInstanceKey({ id: "0x1", className: "TestSchema.TestName" }),
+            createTestInstanceKey({ id: "0x2", className: "TestSchema.TestName" }),
+          ],
+          options: { autoExpand: { depthInPath: 1 } },
+        },
+      ];
+      const filteringFactory = await createFilteringHierarchyDefinition({ nodeIdentifierPaths: paths });
+      const row = {
+        [NodeSelectClauseColumnNames.FullClassName]: "",
+        [ECSQL_COLUMN_NAME_FilterECInstanceId]: "0x2",
+        [ECSQL_COLUMN_NAME_FilterClassName]: "TestSchema.TestName",
+      };
+      const node = await firstValueFrom(filteringFactory.parseNode(row));
       expect(node.autoExpand).to.be.undefined;
     });
 
@@ -486,7 +506,7 @@ describe("FilteringHierarchyDefinition", () => {
       expect(node.autoExpand).to.be.true;
     });
 
-    it("sets auto-expand when all filtered children paths autoExpand depth is greater than parent keys length and includeGroupingNodes is set to true", async () => {
+    it("sets auto-expand when all filtered children paths autoExpand depthInHierarchy is greater than parent keys length", async () => {
       const paths: HierarchyFilteringPath[] = [
         {
           path: [
@@ -494,7 +514,7 @@ describe("FilteringHierarchyDefinition", () => {
             createTestInstanceKey({ id: "0x2", className: "TestSchema.TestName" }),
             createTestInstanceKey({ id: "0x3", className: "TestSchema.TestName" }),
           ],
-          options: { autoExpand: { depth: 3, includeGroupingNodes: true } },
+          options: { autoExpand: { depthInHierarchy: 3 } },
         },
       ];
       const filteringFactory = await createFilteringHierarchyDefinition({ nodeIdentifierPaths: paths });
@@ -862,6 +882,118 @@ describe("FilteringHierarchyDefinition", () => {
         expect(result.autoExpand).to.be.undefined;
       });
 
+      it("doesn't set auto-expand when depthInPath is smaller than grouping node depth", async () => {
+        const groupingNode = createGroupingNode();
+        const inputNode: ProcessedGroupingHierarchyNode = {
+          ...groupingNode,
+          parentKeys: [createTestNodeKey()],
+          children: [
+            {
+              ...createTestProcessedInstanceNode(),
+              filtering: {
+                filteredChildrenIdentifierPaths: [
+                  {
+                    path: [createTestInstanceKey()],
+                    options: {
+                      autoExpand: {
+                        depthInPath: 1,
+                      },
+                    },
+                  },
+                ],
+              },
+            },
+          ],
+        };
+        const filteringFactory = await createFilteringHierarchyDefinition();
+        const result = await firstValueFrom(filteringFactory.postProcessNode(inputNode));
+        expect(result.autoExpand).to.be.undefined;
+      });
+
+      it("sets auto-expand when depthInPath is higher than grouping node depth", async () => {
+        const groupingNode = createGroupingNode();
+        const inputNode: ProcessedGroupingHierarchyNode = {
+          ...groupingNode,
+          parentKeys: [createTestNodeKey()],
+          children: [
+            {
+              ...createTestProcessedInstanceNode(),
+              filtering: {
+                filteredChildrenIdentifierPaths: [
+                  {
+                    path: [createTestInstanceKey()],
+                    options: {
+                      autoExpand: {
+                        depthInPath: 2,
+                      },
+                    },
+                  },
+                ],
+              },
+            },
+          ],
+        };
+        const filteringFactory = await createFilteringHierarchyDefinition();
+        const result = await firstValueFrom(filteringFactory.postProcessNode(inputNode));
+        expect(result.autoExpand).to.be.true;
+      });
+
+      it("doesn't set auto-expand when depthInHierarchy is smaller than grouping node depth", async () => {
+        const groupingNode = createGroupingNode();
+        const inputNode: ProcessedGroupingHierarchyNode = {
+          ...groupingNode,
+          parentKeys: [createTestNodeKey()],
+          children: [
+            {
+              ...createTestProcessedInstanceNode(),
+              filtering: {
+                filteredChildrenIdentifierPaths: [
+                  {
+                    path: [createTestInstanceKey()],
+                    options: {
+                      autoExpand: {
+                        depthInHierarchy: 1,
+                      },
+                    },
+                  },
+                ],
+              },
+            },
+          ],
+        };
+        const filteringFactory = await createFilteringHierarchyDefinition();
+        const result = await firstValueFrom(filteringFactory.postProcessNode(inputNode));
+        expect(result.autoExpand).to.be.undefined;
+      });
+
+      it("sets auto-expand when depthInHierarchy is higher than grouping node depth", async () => {
+        const groupingNode = createGroupingNode();
+        const inputNode: ProcessedGroupingHierarchyNode = {
+          ...groupingNode,
+          parentKeys: [createTestNodeKey()],
+          children: [
+            {
+              ...createTestProcessedInstanceNode(),
+              filtering: {
+                filteredChildrenIdentifierPaths: [
+                  {
+                    path: [createTestInstanceKey()],
+                    options: {
+                      autoExpand: {
+                        depthInHierarchy: 2,
+                      },
+                    },
+                  },
+                ],
+              },
+            },
+          ],
+        };
+        const filteringFactory = await createFilteringHierarchyDefinition();
+        const result = await firstValueFrom(filteringFactory.postProcessNode(inputNode));
+        expect(result.autoExpand).to.be.true;
+      });
+
       it("sets auto-expand when node has hierarchy depth smaller than the filter target", async () => {
         const groupingNode = createGroupingNode();
         const inputNode: ProcessedGroupingHierarchyNode = {
@@ -886,7 +1018,7 @@ describe("FilteringHierarchyDefinition", () => {
         expect(result.autoExpand).to.be.true;
       });
 
-      it("sets auto-expand when node has hierarchy depth smaller than the filter target and same key", async () => {
+      it("sets auto-expand when node has depthInPath smaller than the filter target depth", async () => {
         const groupingNode = createGroupingNode();
         const inputNode: ProcessedGroupingHierarchyNode = {
           ...groupingNode,
@@ -898,7 +1030,7 @@ describe("FilteringHierarchyDefinition", () => {
                 isFilterTarget: true,
                 filterTargetOptions: {
                   autoExpand: {
-                    depth: 2,
+                    depthInPath: 2,
                   },
                 },
               },
@@ -910,7 +1042,7 @@ describe("FilteringHierarchyDefinition", () => {
         expect(result.autoExpand).to.be.true;
       });
 
-      it("sets auto-expand for grouping node whose child has autoExpand depth with includeGroupingNodes set", async function () {
+      it("sets auto-expand when node has depthInHierarchy smaller than the filter target hierarchy depth", async function () {
         const inputNode: ProcessedGroupingHierarchyNode = {
           ...createGroupingNode(),
           parentKeys: [createTestNodeKey()],
@@ -921,55 +1053,7 @@ describe("FilteringHierarchyDefinition", () => {
                 isFilterTarget: true,
                 filterTargetOptions: {
                   autoExpand: {
-                    depth: 2,
-                    includeGroupingNodes: true,
-                  },
-                },
-              },
-            },
-          ],
-        };
-        const filteringFactory = await createFilteringHierarchyDefinition();
-        const result = await firstValueFrom(filteringFactory.postProcessNode(inputNode));
-        expect(result.autoExpand).to.be.true;
-      });
-
-      it("doesn't set auto-expand for grouping node whose child has autoExpand depth pointing to parent with includeGroupingNodes set", async function () {
-        const inputNode: ProcessedGroupingHierarchyNode = {
-          ...createGroupingNode(),
-          parentKeys: [createTestNodeKey(), createGroupingNode().key],
-          children: [
-            {
-              ...createTestProcessedInstanceNode(),
-              filtering: {
-                isFilterTarget: true,
-                filterTargetOptions: {
-                  autoExpand: {
-                    depth: 2,
-                    includeGroupingNodes: true,
-                  },
-                },
-              },
-            },
-          ],
-        };
-        const filteringFactory = await createFilteringHierarchyDefinition();
-        const result = await firstValueFrom(filteringFactory.postProcessNode(inputNode));
-        expect(result.autoExpand).to.be.undefined;
-      });
-
-      it("sets auto-expand for grouping node whose child has autoExpand depth", async function () {
-        const inputNode: ProcessedGroupingHierarchyNode = {
-          ...createGroupingNode(),
-          parentKeys: [createTestNodeKey(), createGroupingNode().key],
-          children: [
-            {
-              ...createTestProcessedInstanceNode(),
-              filtering: {
-                isFilterTarget: true,
-                filterTargetOptions: {
-                  autoExpand: {
-                    depth: 2,
+                    depthInHierarchy: 2,
                   },
                 },
               },