+++ packages/hierarchies/CHANGELOG.md
@@ -38,6 +38,47 @@
 - Updated dependencies:
   - @itwin/presentation-shared@2.0.0-alpha.0
 
+## 1.7.1
+
+### Patch Changes
+
+- [#1024](https://github.com/iTwin/presentation/pull/1024): Update `depthInPath` documentation. Clarified that it does not set auto-expand for node that is at `depthInPath` position.
+
+## 1.7.0
+
+### Minor Changes
+
+- 8f7d200926b93e862276e1f978f6c891691e0dae: Refactored specifying the depth / level to which filtering path should be auto-expanded.
+
+  - Deprecated `FilteringPathAutoExpandOption` and `FilterTargetGroupingNodeInfo`.
+  - Added `FilteringPathAutoExpandDepthInPath` and `FilteringPathAutoExpandDepthInHierarchy` which should be used instead.
+
+  Migration:
+
+  ```ts
+  const pathExpandedToInstanceInPath: HierarchyFilteringPath = {
+    path: [instanceKey1, instanceKey2, instanceKey3],
+    options: {
+      // deprecated:
+      autoExpand: { depth: 1 },
+      // do this instead:
+      autoExpand: { depthInPath: 1 },
+    },
+  };
+
+  const pathExpandedToGroupingNode: HierarchyFilteringPath = {
+    path: [instanceKey1, instanceKey2, instanceKey3],
+    options: {
+      // deprecated:
+      autoExpand: { depth: 1, key: groupingNodeKey },
+      // also deprecated:
+      autoExpand: { depth: 1, includeGroupingNodes: true },
+      // do this instead:
+      autoExpand: { depthInHierarchy: 1 },
+    },
+  };
+  ```
+
 ## 1.6.1
 
 ### Patch Changes